version 1.3.0

+++ Source/Bearing.swift
@@ -11,8 +11,8 @@ import Foundation
 /// Direction in degrees with true north at 0° and progressing clockwise
 public struct Bearing {
     
-    /// `Float` between 0.0 and 1.0 representing 0° and 360°, respectively
-    public let zeroToOne : Float
+    /// `Double` between 0.0 and 1.0 representing 0° and 360°, respectively
+    public let zeroToOne : Double
     
     /// `Int` representing the degrees from true north in a clockwise progression
     public let degrees : Int
@@ -23,8 +23,8 @@ public struct Bearing {
     /// Human-readable cardinal point representation of `Bearing.degrees`
     public let cardinalLabel : String
     
-    init(_ value : Float) {
-        degrees = Int(roundf(value))
+    init(_ value : Double) {
+        degrees = Int(round(value))
         zeroToOne = value / 360
         label = "\(degrees)°"
         let directions : [String] = ["N","NNE","NE","ENE","E","ESE","SE","SSE","S","SSW","SW","WSW","W","WNW","NW","NNW"]
