+++ Source/Percentage.swift
@@ -11,8 +11,8 @@ import Foundation
 /// Contains a value, 0 to 1 representation and label describing percentage
 public struct Percentage {
     
-    /// `Float` between 0.0 and 1.0 representing 0% and 100%, respectively
-    public let zeroToOne : Float
+    /// `Double` between 0.0 and 1.0 representing 0% and 100%, respectively
+    public let zeroToOne : Double
     
     /// `Int` representing a percentage
     public let value : Int
@@ -20,9 +20,9 @@ public struct Percentage {
     /// Human-readable representation of the percentage
     public let label : String
     
-    init(_ value : Float) {
+    init(_ value : Double) {
         zeroToOne = value
-        self.value = Int(roundf(value * 100))
+        self.value = Int(round(value * 100))
         label = "\(self.value)%"
     }
 }
